Added logout button(requires refresh)
