+++ frontend/src/App.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import {BrowserRouter, Route} from "react-router-dom";
 
 // import Home from "./pages/Home";
-import Home from "./pages/HomeClass";
+import Home from "./pages/Home";
 import SignIn from "./pages/SignIn";
 import SignUp from "./pages/SignUp";
 import Navbar from "./pages/components/Navbar";
