+++ frontend/src/pages/Home.tsx
@@ -1,15 +1,41 @@
 import axios from "axios";
-import React, { useState, useEffect } from "react";
+import React, { Component } from "react";
 import "./styles/Home.scss";
+import GetUserResponse from "../responses/GetUserResponse"
+import LogoutButton from "./components/LogoutButton"
 
-const Home = () => {
+type State = {
+  user: GetUserResponse;
+}
+type Props = {
 
-  const [name, setName] = useState('');
+}
 
-  useEffect( () => {
+class Home extends Component<Props, State> {
+  constructor(props: Props) {
+    super(props);
 
-    //get name from response
-    /*
+    this.state = {user: new GetUserResponse('','','')};
+  }
+
+  componentDidMount(){
+    this.getLoginData();
+  }
+
+  render(){
+    return (
+      <section className="home">
+        <h1 className="home-title">Home page</h1>
+        <div>{this.state.user.firstName !== '' ? "Welcome " + this.state.user.firstName : "You are not logged in"}</div>
+        {this.state.user.firstName !== '' && 
+          <LogoutButton />}
+      </section>
+    );
+  }
+
+  getLoginData = () => {
+    /* 
+     * Expected Response Body
      * "id": "3180efe0-7186-4031-87bb-df318617b9a9",
      * "email": "joelstav@outlook.com",
      * "firstName": "Joel"
@@ -19,21 +45,17 @@ const Home = () => {
       headers: { "Content-Type": "application/json" },
       withCredentials: true,
       url: "http://localhost:5000/api/User/oneuser",
-    }).then(response => 
-      console.log(response.data)
-      // this.setState({ nameData: response.data})
-    );
-      // .then(res => setState)
 
-    setName("NON DYNAMIC NAME");
-  });
+    }).then(res => {
+      console.log("Response Data: " + res.data);
+      this.setState({ user: res.data });
+      console.log("FirstName: " + this.state.user.firstName);
 
-  return (
-    <section className="home">
-      <h1 className="home-title">Home page</h1>
-      <div>{name ? "Hi " + name : "You are not logged in"}</div>
-    </section>
-  );
+    }).catch(function (error){
+      console.log("No user login found");
+
+    })
+  }
 }
 
 export default Home;
